Slack notification for tests (#822)
* Add slack integration to tests.

Include the executable github-slack-notify.py file to send notifications to slack. Updates both the daily and hourly tests to use slack notification instead of email.

* Update default notification message and workflow.

This corrects the workflow structure for the daily test by requiring the smoke-test-dev job prior to notification. It also updates incorrect job names to correctly invoke the notification tool. Finally, add a default notification to alert the channel.