Update "compat" funcx-endpoint error imports
In several places, imports are done with an ImportError handler to
manage the relocation of error classes. Now that v0.4.0a2 is published
to pypi, remove this handling.

Equivalent handling is *not* removed from smoke_tests because those
will need to work against the last non-alpha funcx version for now.