Switch to recent parsl master
This commit removes an unnecessary dependency on a
parsl executor superclass that has been restructured
since the previously used parsl release. It turns out
that only one line from that superclass was needed,
to store the specified provider, and this PR moves
that line directly into the FuncX version of HTEX.

This allows a PBS contribution to parsl from Ryan Chard
to be used in FuncX endpoints.

This should also allow more easy updating of the
parsl dependency as/when desired.