Improve forwarding performance with threads
The main loop morally contained two kernels: one to collect and forward tasks
and one to collect and forward results.  The problem was that either we wait
for a time (timeout) or don't block and have a busy loop.  In the worst case of
either only tasks or only results to forward, that meant the timeout for the
"nothing to forward" case puts a bottleneck on the other loop kernel.  At a
`timeout=0.01`, that limits the maximum throughput to 100 iterations per
second.

Address this by pulling the two kernals out into two separate threads, and
hopefully keeping the intent of the heartbeat logging.