Use correct variable: period, not threshold
Nominally a thinko when transitioning the intertwined kernel to multiple
threads: we want a heartbeat every period, not on the threshold of timing out.

In doing so, remove the only usage of threshold, and so also remove the
threshold class variable.  It looks like the last legitimate use of this
variable was removed in Apr:

    commit 08eba06bc1ff365f0b932d9b09bd4c1eb01493f1
    Author: Yadu Nand Babuji <yadudoc1729@gmail.com>
    Date:   Thu Apr 7 15:29:25 2022 -0500

        Integrate rabbitmq with endpoint [RMQ Series #3] (#703)