Correct endpoint shutdown logic (#873)
Previously, the endpoint was shutdown by the cold logic of SIGTERM, wait for
10s, then SIGKILL.  The problem was that the SIGTERM handler was blocking
everytime, and so SIGKILL was always called.

In production, this might be hidden by logs, or ignored because Ctrl+C (a.k.a.,
"SIGINT") is handled differently by default.

The fix: do as *little as possible* in a signal handler.  We now just set a
flag, and act on that flag during our normal processing.