Refactor class variable name
This is a slightly confusing muddling of a variable's purposes in two different
contexts.  In the parent process, the `quiesce` event is utilized to signal a
need to restart ("reload"; "reboot") the communication setup of the funcx
endpoint.  (Shutdown the task queue watcher and result publisher, then restart
them again.)  In the child process, that semantic is lost and the point is just
"time to shutdown."  Meanwhile, while tracing through, the change in name from
`quiesce` to `kill` has proven difficult to follow mentally.  So, choose the
path of (apparent) least resistance and refactor it based on the development
team's preferences.