Introduce delay when quiescing.
If the quiesce event is set, it typically indicates that a connection RMQ has
been faltered somehow.  The fix we have chosen is to restart the
TaskQueueSubscriber and ResultQueueSubscriber subprocesses.  However, if
there's a transient connection issue, restarting too quickly may spam the tubes
too early.  So, instead if restarting ("quiescing"), then wait for 5 seconds.