Remove inappropriate parsl ScalingFailed exception usage
Parsl's ScalingFailed is based around the structure of parsl's
executors and providers scaling mechanism, which is not used in
funcX - that makes ScalingFailed the wrong kind of exception to
raise here.

This commit changes this to a RuntimeException, pending bikeshedding
about proper error types.

A second use of ScalingFailed, also incorrect, exists in what appears
to be dead code in funcx_endpoint/executors/high_throughput/executor.py.
Because that appears to be dead code, this commit does not attempt
to interfere there.