Show error-level logs in tests
After finally, *finally* tracking down the cause of failure in a recent test,
it became clear that the problem could have been identified much sooner if we
had access to the information the _production_ code already handles gracefully.
The test was in error, but it was very subtle what the problem was until the
error-level logs were emitted.  Attempt to avoid that experience going forward.