Colorize console logs
In order to more quickly draw attention to the relevant parts of the log lines,
implement ANSI color escape codes.  The idea is not artistic perfection, but a
simple approach to "syntax highlight" the console (*strictly* console, not log
files).  To that end, the "header" of each line (i.e., time, runtime info, code
line, method name) is the same, but the actual message is slightly different
for DEBUG, INFO, WARNING, and ERROR logs.

In addition, since we are so heavily UUID based, also highlight any UUIDs.

For those who want to log to the console and do not want color, there is also a
`--no-color` argument.