More robustly handle executor failure to start
We can detect most startup errors in the executor interchange fairly quickly:
catch the exception.  Thus, do so, and let the parent process know right away
-- no sense in timing out if we know with certaintly that we're not going to
succeed.  Now the endpoint won't continue but quit immediately -- better
iteration path for the human.

Meanwhile, add a slight delay as a quick-and-dirty hack to avoid interwoven
exception lines from the parent and child processes in the log or output.