Remove unnecessary data structure
The Tasks and associated futures are created at the same time and can always be
co-located.  There's no need to artificially separate them, and then have to
look up one or the other.  Keep them together through the queue, and remove the
unnecessary dict.

Meanwhile, with the change to `.get(block=False)` there's no more need for the
coalesce delay.  Outside of egregious limits (e.g., a *very* high batch_size
or a *very* small time limit), the timeout check is useless.