Serialize access to sqlite
SQLite helpfully defaults `check_same_thread` to `True`, which ensures that all
access to the DB happens on the same thread that created the DB connection.
But that's not *actually* the important bit.  What's important is that all
transactions through that connection with the DB are *serialized*.  Serialized
transaction calls are part of how SQLite guarantees that the underlying tables
keep their integrity -- ensuring that all calls come from the same thread is a
means to that end.  If we externally guarantee that our accesses to the DB are
serialized, then we don't need SQLite to check that access are from the same
thread.

Thus, we ensure that we do this for the important SQLite interactions with an
access lock.