Remove unnecessary use of EP name
Following the thread that the name _is the same as_ the EP configuration
directory, there's need to pass it around.  Since we're already passing the
configuration directory around, just calculate the name when/if needed.

While here, move to some more use of `pathlib`, which engenders a nice
simplification in `register_endpoint` as well.