Add a small amount of typing
The initial change is to the `Interchange.start()` method signature, which then
opens a number of minor items to mypy.  One of those changes is the `log.trace`
method; address by subclassing the logger.  Mypy has learned a little bit since
~a year ago, so the only trickery here is in the `import`:

    from funcx_endpoint.logging_config import FXLogger

    log: FXLogger = logging.getLogger(__name__)  # type: ignore

`logging_config.py` tells the logging system to return FXLogger, but `mypy`
doesn't infer that deeply.  So, promise it that "we know what we're doing," and
then `.trace()` can exist.

The other changes are of the "while here" variety:

 - don't waste time interpolating strings for messages that won't be emitted!

 - cache lookups for easier reading, and to simplify some mypy-isms

 - use if-guard style in a couple of places to reduce indent