Do not pre-emptively test auth status
For 99% of expected use-cases, there is no need to check the authorization
status: the user will have already authed.  Consequently, there is no need to
burden the Auth server with superfluous calls, nor is there a need to add
latency to new FuncXClient instantiations: strictly and _only_ handle
`AuthAPIError`s in the decorator.

Also, while here, reduce the lock scope of the lock.  There's no need to hold
the DB lock while the network call is outstanding.

[sc-20480]