Fine-tune endpoint quiesce interaction
Until this commit, the model has been that if any interruption or exception
happens, we set the quiesce flag, shutdown, and then restart communication.
That has included registering the endpoint.

As of this commit, that logic changes to expect that an endpoint registration
happens only once -- at start -- and any internal quiesce actions only restart
communication to the AMQP service.

Accordingly, re-integrate the `register_endpoint` logic into the
`Endpoint.start_endpoint()` method and update tests.

One key outcome is that the Interchange class is no longer concerned with *how*
to get AMQP credentials -- it requires that they be provided at instantiation.

[sc-19623]