Remove use of FuncXClient from interchanges
The last remaining use of the FuncXClient was within the *executor*
interchange, to enable looking up container location information.  The required
location information is now passed along with each task message, so modify the
logic to look there -- the key change is in `messages_compat.py`.  Everything
else is just "cleanup."

[sc-19623]