Teach FuncXExecutor to use container_id natively
The ability for the FuncXExecutor to use containers was not *completely* lost
in the recent FuncXExecutor refactor (81811a97e0), but was not enabled in the
default workflow.  In particular, the expectation from the refactor is that the
`endpoint_id` and `container_id` are specified as FuncXExecutor attributes, and
i_not_ as part of `.submit()`.  But that implies that `.submit()` will
transparently use them -- it mistakenly did not.

Previously, one would have had to use `.register_function()` manually and know
to specifically specify the `container_uuid` kwarg (eww!), or manually
associated the function with a container via the FuncXClient methods.  Further,
the same function could only be associated with a single container.  Addressed
by updating the internal function registry to also take into account the
`container_id`.

[sc-21526]