Specifically handle the user's umask
In the case that a user has an unexpected umask, we now fall in line with the
expectation of "still working."  In particular, we follow the main thread of
how GNU's `mkdir` handles permissions, and force bits 2 and 1, but let the user
still specify bit 3 (read).  So, if a user has a umask like 0o0777 (which
creates files with no permissions), then we honor the bit at 0o0400, resulting
in directory permissions of:

    d-wx------

The other change of note is now toward slightly more secure by default: default
the .funcx/ permissions to be owner-accessible only.  No group or other access.

[sc-22042]