Implement flag to maintain process tree
We want to be able to force -- despite a misconfiguration -- that an
administrator can see a hierarchical view of all Globus Compute endpoints.  In
particular, if the parent process crashes, the child processes should
gracefully shutdown rather than hang around as orphans.  Thus, implement
`--die-with-parent` as a flag to the cli to indicate that if the parent PID
changes -- so the child process was reparented or the parent process died --
then it is time to shutdown.