Implement EndpointManager
The `EndpointManager` connects to the Globus Compute web services to register
as a multi-tenant endpoint, and then listens for commands from upstream via an
AMQP queue.

Commands are verified via:

- Command names must follow the regex `^[A-Za-z][0-9A-Za-z_]{0,99}$`

    - Roughly, a valid, non-private Python method name

- Commands must have an implemented class method on `EndpointManager`

- Commands must be associated with a Globus username

- The local machine must have a mapping of the Globus username to the local
  username

After validation (the above 4 bullet points), the command is then forked to a
child process, and that child process drops all privileges by becoming the
local user (i.e., `os.setres*id()`)

After forking, child processes are not double-forked before execing, with the
intent of maintaining the process hierarchy tree for easier administrative
understanding and maintenance.  Similarly, child processes are signalled to
shutdown (SIGTERM) if the parent process shuts down.  After a 10 second grace
period, the child processes are sent the SIGKILL method.  Notably, signals are
sent strictly from the same user as which owns the child processes -- the
parent process user does *not* send the signals.

As of this commit, the only implemented command is `start_endpoint`.

[sc-19623]