Prefix commands methods with `cmd_`
Coming out of a post-merge review and walk-through, there was some fear that
just ensuring the method existed on the class and was not private was not
enough of a safe-guard, API wise.  An additional clue to the next dev when
creating new functions is that they must now be prefixed with `cmd_`, which
means the dev is aware of what they're implementing, and it will be much, much
more difficult for a rogue execution to slip through.

Note that this is *in addition* to the web-side validation, which includes a
specific allow-list (rather than a catch-all, like a regular expression).

[sc-19623]