Re-allow endpoints to instantiate credentials
The fix is simple: remove the errantly added five lines in endpoint.py.
However, testing this particular scenario proved difficult in standard pytest,
so instead, opt for an outside-of-pytest approach: simple shell scripts.

This is not a complete solution for non-Pytest tests -- in particular, there's
currently no distinction between prod and dev or local -- but there's also not
a current need for it.  Assume we'll update the runner.sh (and tests) as/when
the need is obviated.

[sc-23429]