Pass GC_TASK_TIMEOUT to instantianted pod
The helm chart faithfully exports the `GC_TASK_TIMEOUT` variable to the
endpoint's environment, but the workers weren't receiving that variable.

The implemented fix is a hard-coded list of environment variables (okay, a list
of length 1) to check.  If passing information from the parent pod (the
endpoint) to the child pods (the workers) becomes an issue in the future, the
intended approach is to "deal with it then" in a more complete way.  For now,
then, though not ideal, this commit is "good enough."

[sc-23725]