Send EPStatusReport even if no outstanding tasks
The interchange should send an EPStatusReport every `heartbeat_period`, period.
273e39002cf1ceda613ba6ad7893b87549c22046 broke that expectation for the case of
"no task deltas" -- a common occurrence for the tests we run because when a
task is done, it's delta is handled directly by the task-specific result
message and removed from the outstanding "task status deltas" data structure.

Even more than tests, however, it's also a problem when there are actually no
tasks running -- when the endpoint is legitimately idle, it should still be
marked as "online."

Follow up for [sc-19273]