Implement support for YAML config
We now store configuration variables in a config.yaml file and
dynamically generate a `Config` object using the file's content.

For backward compatibility, we also support manually specifying the
config in a config.py file. We will ignore the config.yaml file when a
config.py file is in the endpoint dir.