Add user-facing EP name validation
We recently implemented provisions for users to have alternate _display names_
(via `--display-name` to the `configure` subcommand, or to manually alter the
`config.py` top-level configuration item `display_name="some name"`), but
didn't go the next step and enforce that users can't do something silly (but
natural!) like adding spaces to the name.  This has repercussions later on in
that Parsl is not very robust with it's path handling.  So, assume the better
part of valor is to die early with an explanation for the user rather than
cryptically failing later on with a traceback.

[sc-24521]