Fix broken test (#1208)
The test was utilizing the `fs` mock and also `patch()`ing a method.  The patch
returned a `pathlib.Path` object, but was applied _before_ the `fs` fixture had
setup.  Update test so that the `fs` fixture is in place prior to the mock.

Meanwhile, after the test was corrected, noted that the logic under test was
not completely correct per the method documentation: replaced `str.replace`
with `re.sub("^...")` so as to match strictly at the beginning of lines.

While here, apply a couple of other minor refactors:

- add `funcx_endpoint.engines` to what is considered for replacement (per the
  recently merged Engine work)
- use _MOCK_BASE to simplify some logic in the test file
- remove used test parameters (always `False`)
- correct fallback test logic (`assert AssertionError` &rarr; `raise AssertionError`)
- refactor `funcx` to `compute` in `test_endpoint.py`