Address race-condition w/ Executor task submission (#1248)
Prior to this commit, tasks were not associated with a taskgroup.  In the case
that a user wanted to switch task groups mid-run this created a race condition:

```python
with Executor(...) as ex:
  ex.task_group_id = "some-task-group-id"
  ex.submit(some_func_01)
  ex.task_group_id = "some-other-task-group-id"
  ex.submit(some_func_02)
```

Until this commit, `.submit()` calls did not associate tasks with a task group
id, instead creating the association _with the currently assigned Executor
`task_group_id`_ when the internal thread pulled the tasks for upstream
submission.  Thus, from the above code, there was no guarantee that
`some_func_01` would be associated with `some-task-group-id`.

Address by associating each task with the current task_group_id before
returning the future to the calling code.

[sc-25897]