Store the task group as a native uuid (#1246)
Remove downstream need to handle both cases.  It's now always a UUID.

Unfortunately, due to a long-open bug,[1] mypy does not know how to handle data
normalization via the properties.  A workaround appears to be to explicitly
state the private variable's type, rather than specify the type on the property
directly.

[1] https://github.com/python/mypy/issues/3004
  "What to do about setters of a different type than their property?"