Address slow test
Test was taking close to 18s on my machine, due to useless and idle waiting,
and then producing a traceback.  The traceback wasn't important to the test,
but was noise "of concern" until I figured it out.  Realized that for the
original intent of the test (written in mid 2021), there was no need to start
up a child process.  Now uses the `ThreadPoolEngine` for a much faster
turnaround time with no traceback.