Remove unpack() in endpoint/interchange (#1252)
One of the main goals of the recent Engine work and the addition of the task
identifying information to the AMQP messages was to remove the need to
deserialize in the main endpoint interchange.  More recently,[1] we removed the
`unpack()` call from the Task forwarding thread loop.  Follow-through and
remove the same from the Result forwarding thread loop as well.

[1] 1ee3fd54906895393a78a6916f19edef70a31c76
    "check rmq message header for allowlist and add header to queue (#1241)"