Address cross-thread access to Pika (#1255)
This commit largely reimplements the Result Publisher logic, now utilizing
Pika's `SelectConnection` setup for communicating over AMQP.  In so doing, the
result publisher now exists in its own thread, and messages published to it are
enqueued rather than immediately published to the wire.  Consequently, the
biggest external change is that `.publish(msg)` now returns a `Future[None]`,
that can be queried to determine when server has acknowledged receipt.

This implementation has the knock-on effect of being vastly more performant.

[sc-26048]