Prevent same-user child-processes (#1266)
For administrators running a multi-tenant endpoint, the parent process -- the
endpoint manager -- is expected to lookup local user ids (UID) and then drop
privileges to that UID before `exec()`ing.  On the other hand, for non-power
users (e.g., not `root`) running the multi-tenant endpoint, the process won't
have the privileges to change users anyway.  Consequently, implement a check
that only allows the parent endpoint process to create child endpoint processes
("user endpoints") with the same process UID if the UID is not otherwise
privileged.

This is implemented with 2 checks:

- Is the parent process owned by UID 0?
- Does the parent process have any admin-like capabilities (via `pyprctl`)

If either are true, then set a flag to prevent any attempts to create a child
process with the same UID.

If neither are true -- so, when a "regular user" is running the multitenant
endpoint manager -- then only proceed if the MT UID and the local user UID
_do_, in fact, match -- don't even try to execute the `os.set*id()` syscalls.

[sc-26381]