Allow macOS to test
For MVP, we don't support MT on macOS.  Right now, the APIs we utilized are not
supported on macOS (e.g., `os.setresuid()`) and so are not expected to run.
That went further recently with the introduction of `pyprctl` which uses ctypes
to make `prctl` syscalls to the kernel.  However, the macOS privilege model is
different, and there is no `prctl` exported by the kernel.

Quick fix for now then: don't run the tests on macOS.