Remove noisy log during test
Due to `click.CliRunner`'s implementation for testing sys.stdin, and our use of
sys.stdin for credentials, we're hitting the early `log.debug()` call and
getting:

    ... Invalid info on stdin -- (JSONDecodeError) Expecting value: line 1 column 1 (char 0)

when showing the verbose debugging output.  This is noise that is unrelated to
the tests, so opt instead to give a default value that avoids that log call.