Quiet some test output
While en route to tracking down a test failure in another branch (only
occurring in CI -- doh!), was overwhelmed with a number of superfluous output
generated by some tests.  Realized that we don't see this locally because we
have not had parity with CI `--log-cli-level=ERROR` flag to `pytest`.  Changing
that (`tox.ini`), and then clearing up the then-apparent noisy tests.

As a side-benefit, some of these noisy tests weren't actually testing what they
were supposed to, and the error messages were, in fact, telling us so.  So ...
"fixed."

Take the opportunity to update mypy