Fix two assertions
The `assert` statement merely checks the truthiness of its operand.
Unfortunately, that's a problem when there's a typo, as was the case here.
`MagicMock().called` returns a boolean of whether that attribute was called as
a function, but `MagicMock().called_with(...)` is not defined specially by
`unittest.mock`.  Thus, it is magically being created by that very statement,
and just passing.

What _does_ exist is `.assert_called_with(...)`, so alter tests accordingly.