Read template before dropping privs (#1360)
The error was in attempting to read a file to which the dropped-privileges
process did not have access:

    /root/

`root`'s home directory is (very correctly) set to 0o40700, meaning that only
`root` can access anything inside of `/root/`.  This is a problem for the user
endpoint as it is unable to read the `user_config_template.yaml` file that is
located at

    /root/.globus_compute/<multi_user_endpoint>/user_config_template.yaml

Addressed by reading the file into a string *before* dropping privileges.

[sc-28360]