First go-round at MU instructions (#1368)
This documentation is mostly oriented around a "tl;dnr" section to get a
multi-user endpoint admin up and running quickly.  For the identity mapping
logic, however, I have not yet figured out how to succinctly put that, so that
points to the GCS documentation, and then directs folks to reach-out in help.

Finally, this includes some background information, including the rough flow of
a task from the SDK to the UEP, and where the MEP sits.

[sc-28145]