Disallow an open-endpoint, except by root (#1375)
The identity mapping logic is a user-defined mapping from globus identities to
anything.  That's intended to be for administrators to be able to
programmatically map to organization specific users (for example,
`kevin@example.com` -> local user `kevin` -> local uid `12345`).  However, this
could also be abused to map _all_ identities to the same user, making,
effectively, an open endpoint.  If the administrator does this, okay, but we do
not want a non-power user to be able to create an open endpoint.

The logic implemented here is to not instantiate an identity mapper (i.e.,
ignore any configuration, if specified) if the user running the MU endpoint
does not have user-changing capabilities.  (i.e., is not `root` or does not
have the SETUID privilege.)  Further, the MEP will ignore any requests from
Globus identities that do not match the identity of the Globus user that
authenticated the MEP.

[sc-28396]