Executor: rename funcx_client arg to client
This is achieved by adding a new client argument that has higher
priority than the existing funcx_client argument, and by emitting a
warning to users if they supply funcx_client.

Also standardize on the usage of '' vs `` in the Executor deprecation
warning messages ('' is used by warnings in other classes).