Remove detritus from CI image (#1384)
Safety is pedantically correct, if overzealous, and the provided image in
ubuntu-latest contains some extra pre-installed packages.  This appears to
amount to a situation where pip does not properly remove an old version of an
updated package.  In this case, `pip` itself.  So ... help it along by manually
removing the offending package.

Hopefully, this will become a non-issue shortly once the image is updated again
-- a transient error.  Meanwhile, we have a new CI cruft of our own (`rm -rf
...`)