Be precise about repr() in exceptions
Be precise about intention to use `repr()` in the message/exception formats.
These are NFC statement changes because `__str__` is not currently defined.
(Should it ever become so, "now we're prepared.")

While here, more fully match the Executor specification to raise a
`RuntimeError` for `.map()` as well, if called after shutdown.