Update parsl pin to 2023.12.4 (#1385)
There are a bunch of fixes and updates to the HighThroughputExecutor
on the Parsl side that we are not using because the endpoint is pinned
to parsl==2023.7.3. This PR updates the Parsl version requirements to
parsl==2023.12.4, and makes adds a minor fix to an updated import path
on the parsl side.