Add warning for stuck thread and early stop
Recreating case:

```python
with Executor(
    endpoint_id, user_endpoint_config={"endpoint_setup": "something that will break"}
) as ex:
    ex.submit(noop).result(2)
```

The web service (currently) takes about 30s to give up waiting for the UEP to
come online.  But the code will only wait for 2s (`.result()`).
Unfortunately, if the underlying HTTP request to the API is outstanding, there
is no way to interrupt it.

So, don't try too hard.  If the script continues, then it will _eventually_
time out, but if the script is about to shutdown, let it (via `daemon=True`).
There's no data the user cares about in the thread at that point.  Meanwhile,
at least mention it (`log.warning`) for transparency.

[sc-28608]