Fix flubbed default user template (#1392)
As identified in the associated ticket, there is no `endpoint_init`, and the
`endpoint_setup` item should be top-level.

However, debugging this shouldn't have been quite as arcane as it was: there
were no logs.  In lieu of an upcoming commit to more robustly communicate UEP
errors via the already-instantiated AMQP credentials, this commit ensures that
at least something gets written to the logs in the identified case.

[sc-28607]