Be more transparent with identity mapping errors (#1394)
If an identity fails to map with a particular mapper, the log output can by
slightly confusing.  In particular, there's no indication of which mapper is
failing with which identity.  Improve that situation with the index of each
mapper, the mapper class, the index of the identity (which is, admittedly,
arbitrary, but at least shows that the tested identity is *different* from the
previous iteration), and the sub of the identity.

Old output looked like:

    2023-12-15 [...]:212 map_identity Identity mapper failed -- (IdentityMappingError) Error mapping identity

New output looks like:

    2023-12-15 [...]:213 map_identity Identity mapper failed for mapper 1 (ExternalIdentityMapping) with identity 2 (abc91c2e-1234-4d38-9a8f-1cbf4a6ce811) -- (IdentityMappingError) Error mapping identity

There's further work to go to learn *what* happened, but that needs to happen
in the library.