Fix GCEngine system failure test
The `scale_in` method is called too soon after `submit` in this test,
not giving enough time for the manager to start up and report itself to
the interchange. We will instead kill the manager via a function sent to
the worker.

The test appeared to work because we set the `heartbeat_period` and
`heartbeat_threshold`` to 1 second. The actual heartbeat will almost
always take slightly more time than the configured `heartbeat_period`,
which leads to the interchange marking the manager as lost.