Simplify log formatting
Rather than a series of if-statements in multiple spots, opt for a
`defaultdict` with a pre-defined set of formatters for common four logging
levels.  This makes selection much easier (`formatter =
formatters[record.levelno]`).  At the same time, don't attempt to find each
individual case of uuid for highlighting.  Instead, format the record to a log
line (i.e., a string), and search/replace the string exactly once.  Much
simpler.