Avoid potential MEP shutdown hang
When sending the final heartbeat, do not assume that the heartbeat will ever
actually send.  Instead, give it an upper bound of 10 seconds (arbitrarily
chosen).  This avoids the real case of the endpoint shutting down due to an RMQ
connection issue (true AWS story!) but the result-publisher thread shutting
down after having accepted to send the final HB but before it was sent.