Don't assume UEP fork() can access MEP fds (#1475)
After `fork()`ing, the UEP's logger is still setup to write to the MEP's
configuration.  If the MEP's logging file is inaccessible to the UEP user (as
it should be!), then the UEP's logging attempts will fail.  This is corrected
after the `exec()` call as all file descriptors are closed, and the new process
takes over thereafter, but the intervening `log.debug()` calls produce an
error.  For those intervening calls, then, unset the logfile entirely in the
child process before continuing.

Why didn't we see this before now?  A combination of factors, including what
the logging module does to be smart.  The combination of events:

- a non-seekable `endpoint.log` (e.g., symlink to `/dev/null`)

- we hadn't used `--debug` with the above

- RotatingFileHandler doing the right thing and not attempting to rotate
  non-standard files.

[sc-30480]