Globally address non-seekable log file
PR #1475 addressed the case of the UEP incorrectly assuming grandfathered
access to the MEP's log file descriptor.  A tangential issue, however, is that
some files are not rotatable.  For example, it is entirely feasible to symlink
`endpoint.log` to `/dev/stdout`, in which case the RFH does the right thing and
ignores it -- but it shouldn't have been setup in the first place.  So, do the
first place thing, and don't set up the more correct logging handler.

[sc-30480]