Reset retry counter after 2 HBs
The EP world roughly revolves around heartbeats, an "I'm alive" notification
that goes out, by default, every 30s.  In the interchange, we also use this
interval to determine whether our connection is stable.  Unfortunately, if we
only detect that our connection is stable _when we attempt to send a
heartbeat_, then it's anyone's guess which thread wins: "successfully alive for
a HB, mark as stable" vs "quiesce because we're clearly not stable."

This manifests as an EP staying up for much longer than the settings would
otherwise attempt to enforce -- in the motivating case, almost an hour before
the race condition finally caught up to the settings which wanted the EP to
shutdown in roughly 2m, preventing the fix to the larger errant condition.

Work around this for now by bumping the heuristic of "still alive ==> stable",
to "alive for 2 HBs ==> stable".  That should give plenty of time for an actual
"integration test" (of sorts) to verify that all the pipes are in working
order.

[sc-31352]