Remove now-stale code
We rebranded last April.  We removed the shim in January.  Now after a year,
it's time to just fail and let users install Globus Compute proper.

Meanwhile, since the shim is no longer available, and this import exception
handing code will no longer be in place, also remove the now-stale
documentation, and instead point to the last-pertinent release version of it:
2.18.1