Snapshot per-task uep submission configuration
We've discussed the ability for users to modify the UEP configuration at
submission time, but we've not been explicit that they need create whole new
objects when doing so.  So, do what we should have done in the first place, and
snapshot the TaskSubmissionInfo.

Notably, this does _not_ address the possibility that users modify the args or
kwargs to a function, but saving that for a future moment as that potentially
comes with some memory concerns.  (This does too, but I think will be minimal
compared to the possibility that args/kwargs might be much larger.)

[sc-33284]