Fix broken MPI test
As evidenced by a recent CI failure, the tests expecta `nodeslist` to be a list
of least 1 item in length.  As written, there's 1 in 99 change that it doesn't:

```python
limit = random.randint(1, 100)  # but say limit yields 1, then:
yield [f"NODE{node_i}" for node_i in range(1, limit)]  # is an empty list.  DOH!
```