Fix test-passed data
The underlying tested code was not incorrect, but the live call log was
reporting a suspicious traceback, culminating in a Pydantic error:

    pydantic.v1.error_wrappers.ValidationError: 1 validation error for Result
    task_id
      value is not a valid uuid (type=type_error.uuid)

That was caused by incorrect data sent by the test, resulting in a task_id of:

    ...
    'task_id': <function double at 0x71221d039d00>
    ...

Clearly incorrect.  So, correct the test call-site to more strongly verify that
the test tests the appropriate detail and to remove the suspicious traceback.
The implementation is mainly fixing `test_proc_pool_engine_not_started`, but
update the surrounding tests in the module with the newly implemented fixtures
for consistency.