Move conf debug emission to preexec
The pre-exec environment was the initial place for the configuration emission
during development, but that proved a problem as it was never stored -- it
wasn't very debuggable for regular users!  Thus, moved it to the post-exec
environment so it would show up in logs.

However, this poses a problem if *something* happens before logging is setup
and the string is written -- debugging the rendered template is not possible
then.  Instead, piggy back on the recent std-stream to `endpoint.log` work to
move configuration emission to prior to exec, and guarantee that it makes it to
the log file.  This also avoids the odd check-if-config-is-a-string to decide
whether to write it to the log.  The pre-exec environment has the information
that it's ephemeral and can thus just log it.