Rewrite MEP doc as additional, not either/or
The first round of MEP documentation focused on the intended audience of the
MEP --- administrators.  But the MEP is *actually* a glorified templating
system that even non-administrators can utilize.  So, mention the templating
capabilities in the "normal" endpoint documentation, then rework the
MEP-specific documentation with a focus on administrators.

- Also declutter the Endpoint configuration examples by moving the
  GlobusComputeEngine discussion to the main endpoints page

- Incorporate the overview into the main Endpoint documentation page --- no
  longer hide it in the oddly-clickable parent-menu item.

- Reworked and moved the Security Posture section (MEP page)

- Move installation instructions to a dedicated page

In short, this is a CEA/MEP documentation refactor.

[sc-35263]