Don't test Parsl; test GC code
Starting as an initial warning fix (`max_workers`), refactor tests to verify
the Compute additions independently of Parsl -- Compute is not Parsl so let
Parsl be responsible for its own testing.  In so doing, reduce the time for
each test and more strongly assert correctness via use of `randomstring`.  Move
updated tests to `tests/unit/`.