Add GlobusApp support to Client and WebClient
The SDK `Client` and `WebClient` now support using a `GlobusApp` for
authentication. The `Client` defaults to using a `GlobusApp` if its `app`
argument is `None`. A custom `LoginManager` can still be used for the
`Client`, but it cannot be used together with a `GlobusApp`.

This work also involved migrating from the Globus SDK's `SQLiteAdapter`
class to `SQLiteTokenStorage`. The main difference between the two is
that the former creates an additional `config_storage` table, which
we do not utilize.