Teach Executor a bursty rate-limit
The Executor currently drains the incoming task queue or stops draining at
`.batch_size` before submitting to the API.  If tasks are `.submit()`ed at just
the right cadence, it's possible to get the Executor to send lots of small
batches very frequently rather than coalescing more tasks.

The fix implemented here is to allow up to 4 back-to-back API `/submit` calls
within a window of 16s _of submissions that do not fully utilize the entire
batch_.  If more API calls come in after than, then induce a 4s delay so as to
play nice with the API.  The burst window will hopefully be ample to work
within a Jupyter notebook and not be an annoyance to a human working through
cells, and only a 4s delay after that until the window clears.  Meanwhile, for
power users, the burst limit and window are configurable (to a point).

Note that this is mostly intended as an education to users -- this is
client-side and so completely up to the user to ignore if they so choose.  One
example: simply create a new Executor -- the state is entirely kept within a
single executor so creating a new instance would be a simple workaround.

(Discovered via log trawling and conversations with a user in funcx#help.)

[sc-35759]